refactor out utils, experiment with getting highlightIndex working
